Populates store at search
