+++ src/App.vue
@@ -23,7 +23,6 @@
 
 <script>
 import Search from './components/Search'
-import axios from "axios";
 
 export default {
   name: 'App',
@@ -35,24 +34,6 @@ export default {
       //
     }
   },
-  mounted() {
-    const data = {key:"free",id:"9m9c8U4f",data:{search:"hello"}};
-    const headers = { 
-        "X-RapidAPI-Host": "macgyverapi-music-graph-v1.p.rapidapi.com",
-        "Content-Type": "application/json",
-        "X-RapidAPI-Key": "173c4655f1mshb630819376b0c16p1f3c16jsnd176b234b6eb"
-      };
-    axios.post("https://macgyverapi-music-graph-v1.p.rapidapi.com/", data, {headers: headers})
-      .then(result => {
-       this.$store.commit("setSongList", result.data.result);
-      })
-  },
-  computed: {
-    songList() {
-      return this.$store.getters.songList;
-    }
-  }
-
 }
 </script>
