+++ src/components/Search.vue
@@ -14,7 +14,7 @@
         mb-5
         xs12
       >
-        <v-form>
+        <v-form @submit.prevent="fetchSongList">
           <v-container>
             <v-layout row wrap justify-center>
               <v-flex xs12 sm6 md30>
@@ -22,6 +22,7 @@
                   label="Song Title"
                   append-icon="search"
                   solo
+                  v-model="searchTerm"
                 ></v-text-field>
               </v-flex>
             </v-layout>
@@ -38,60 +39,25 @@
 </template>
 
 <script>
+  import axios from "axios"
   export default {
     data: () => ({
-      ecosystem: [
-        {
-          text: 'vuetify-loader',
-          href: 'https://github.com/vuetifyjs/vuetify-loader'
-        },
-        {
-          text: 'github',
-          href: 'https://github.com/vuetifyjs/vuetify'
-        },
-        {
-          text: 'awesome-vuetify',
-          href: 'https://github.com/vuetifyjs/awesome-vuetify'
-        }
-      ],
-      importantLinks: [
-        {
-          text: 'Documentation',
-          href: 'https://vuetifyjs.com'
-        },
-        {
-          text: 'Chat',
-          href: 'https://community.vuetifyjs.com'
-        },
-        {
-          text: 'Made with Vuetify',
-          href: 'https://madewithvuetifyjs.com'
-        },
-        {
-          text: 'Twitter',
-          href: 'https://twitter.com/vuetifyjs'
-        },
-        {
-          text: 'Articles',
-          href: 'https://medium.com/vuetify'
-        }
-      ],
-      whatsNext: [
-        {
-          text: 'Explore components',
-          href: 'https://vuetifyjs.com/components/api-explorer'
-        },
-        {
-          text: 'Select a layout',
-          href: 'https://vuetifyjs.com/layout/pre-defined'
-        },
-        {
-          text: 'Frequently Asked Questions',
-          href: 'https://vuetifyjs.com/getting-started/frequently-asked-questions'
-        }
-
-      ]
-    })
+      searchTerm: ""
+    }),
+    methods: {
+    fetchSongList(){
+      const data = {key:"free",id:"9m9c8U4f",data:{search:this.searchTerm}};
+      const headers = { 
+          "X-RapidAPI-Host": "macgyverapi-music-graph-v1.p.rapidapi.com",
+          "Content-Type": "application/json",
+          "X-RapidAPI-Key": "173c4655f1mshb630819376b0c16p1f3c16jsnd176b234b6eb"
+        };
+      axios.post("https://macgyverapi-music-graph-v1.p.rapidapi.com/", data, {headers: headers})
+        .then(result => {
+        this.$store.commit("setSongList", result.data.result);
+        })
+      }
+    },
   }
 </script>
